refactor(internal/database): move schema file to postgres package
