+++ internal/database/metadata/postgres/create_database.go
@@ -35,29 +35,29 @@ func createMetaSchemas(ctx context.Context, opt types.PostgresDbOpt) (err error)
 	// on the same connection: http://go-database-sql.org/modifying.html
 	return database.WithTransaction(db.DB, ctx, func(ctx context.Context, tx *sqlx.Tx) error {
 		// create database functions
-		for _, fn := range database.DB_FUNCTIONS {
+		for _, fn := range DB_FUNCTIONS {
 			if _, err = tx.ExecContext(ctx, fn); err != nil {
 				return err
 			}
 		}
 
 		// create meta tables
-		for _, schema := range database.META_TABLE_SCHEMAS {
+		for _, schema := range META_TABLE_SCHEMAS {
 			if _, err = tx.ExecContext(ctx, schema); err != nil {
 				return err
 			}
 		}
 
 		// create meta views
-		for _, schema := range database.META_VIEW_SCHEMAS {
+		for _, schema := range META_VIEW_SCHEMAS {
 			if _, err = tx.ExecContext(ctx, schema); err != nil {
 				return err
 			}
 		}
 
 		// create triggers
-		for table := range database.META_TABLE_SCHEMAS {
-			trigger := strings.ReplaceAll(database.TRIGGER_TEMPLATE, `{{TABLE_NAME}}`, table)
+		for table := range META_TABLE_SCHEMAS {
+			trigger := strings.ReplaceAll(TRIGGER_TEMPLATE, `{{TABLE_NAME}}`, table)
 			if _, err = tx.ExecContext(ctx, trigger); err != nil {
 				return err
 			}
