+++ internal/database/metadata/postgres/schema.go
@@ -1,4 +1,4 @@
-package database
+package postgres
 
 var DB_FUNCTIONS = []string{
 	`
